% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_corporate_entities.R
\name{agents_corporate_entities}
\alias{agents_corporate_entities}
\title{Agents - Corporate entities}
\usage{
agents_corporate_entities(page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Agents - Corporate entities
}
\examples{
\dontrun{
agents_corporate_entities()
agents_corporate_entities(limit = 3)
agents_corporate_entities(limit = 3, parse = FALSE)
}
}

