% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_families.R
\name{agents_families}
\alias{agents_families}
\title{Agents - families}
\usage{
agents_families(id = NULL, all_ids = FALSE, page = 1, limit = 10,
  parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) a single family agent identifier. optional}

\item{all_ids}{(logical) Return a list of all object ids.
Not sure this does anything? Default: \code{FALSE}}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Agents - families
}
\examples{
\dontrun{
agents_families()
agents_families(id = 3)
agents_families(id = c(6, 7))
agents_families(limit = 2)
agents_families(limit = 2, parse = FALSE)
}
}

