% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_people.R
\name{agents_people}
\alias{agents_people}
\title{Agents - people}
\usage{
agents_people(id = NULL, page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) family agent identifier. optional}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Agents - people
}
\examples{
\dontrun{
agents_people()
agents_people(id = 6)
agents_people(id = 6, parse = FALSE)
agents_people(limit = 2)
agents_people(limit = 2, parse = FALSE)
}
}

