% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_software.R
\name{agents_software}
\alias{agents_software}
\title{Agents - software}
\usage{
agents_software(id = NULL, page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) family agent identifier. optional}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Agents - software
}
\examples{
\dontrun{
agents_software()
agents_software(id = 1)
agents_software(parse = FALSE)
}
}

