% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspacer_search.R
\name{asp_search}
\alias{asp_search}
\title{Search an ArchiveSpace API}
\usage{
asp_search(query, page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{query}{(character) query string, required}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\description{
Search an ArchiveSpace API
}
\examples{
\dontrun{
asp_search(query = 'ecology')
asp_search(query = 'ecology', parse = FALSE)
asp_search(query = 'ecology', limit = 2)
}
}

