% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp_user.R
\name{asp_user}
\alias{asp_user}
\title{Get currently logged in user}
\usage{
asp_user(parse = TRUE, ...)
}
\arguments{
\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\description{
Get currently logged in user
}
\examples{
\dontrun{
asp_user()
}
}

