% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspacer-package.R
\docType{package}
\name{aspacer-package}
\alias{aspacer}
\alias{aspacer-package}
\title{aspacer - ArchiveSpace R client}
\description{
aspacer - ArchiveSpace R client
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

