% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp_login.R
\name{auth}
\alias{asp_login}
\alias{asp_logout}
\alias{auth}
\title{Login to or logout from an ArchiveSpace API}
\usage{
asp_login(user = NULL, pwd = NULL, ...)

asp_logout(...)
}
\arguments{
\item{user}{(character) username, required}

\item{pwd}{(character) password, required}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\description{
Login to or logout from an ArchiveSpace API
}
\details{
retrieves a session ID, expires in 1 hr. the session ID is cached
during your R session, and is gone once you restart the session - thus, you'll
have to login again if you restart your R session
}
\examples{
\dontrun{
# asp_login()
# asp_logout()
}
}

