% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_enumerations.R
\name{config_enumerations}
\alias{config_enumerations}
\title{Config enumerations}
\usage{
config_enumerations(id = NULL, page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) family agent identifier. optional}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Config enumerations
}
\examples{
\dontrun{
config_enumerations()
config_enumerations(id = 3)
config_enumerations(id = 3, parse = FALSE)

library("httr")
x <- config_enumerations(config = progress())
x <- config_enumerations(config = verbose())
}
}

