% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefs.R
\name{prefs}
\alias{prefs}
\title{Get global preferences for the current logged in user}
\usage{
prefs(parse = TRUE, ...)
}
\arguments{
\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Get global preferences for the current logged in user
}
\examples{
\dontrun{
prefs()
}
}

