% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{repos}
\alias{repos}
\alias{repos_accessions}
\title{Get repositories}
\usage{
repos(id = NULL, parse = TRUE, ...)

repos_accessions(repo_id = NULL, page = 1, limit = 10, parse = TRUE,
  ...)
}
\arguments{
\item{id}{(integer) location identifier. optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}

\item{repo_id}{(integer) repository identifier. required}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}
}
\value{
either a data.frame or a list
}
\description{
Get repositories
}
\examples{
\dontrun{
repos()
repos(id = 1)

# doesn't work
# repos_accessions(1)
}
}

