% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories.R
\name{repositories}
\alias{repositories}
\title{Get repositories}
\usage{
repositories(id = NULL, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) location identifier. optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Get repositories
}
\examples{
\dontrun{
repositories()
repositories(id = 1)
}
}

