% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repositories_accessions.R
\name{repositories_accessions}
\alias{repositories_accessions}
\title{Get repositories accessions}
\usage{
repositories_accessions(id, page = 1, limit = 10, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer) location identifier. optional}

\item{page}{(character) page number (default: 1), optional}

\item{limit}{(character) records to return (default: 10, minimum: 10), optional}

\item{parse}{(logical) Attempt to parse to data.frame's if possible. Default: \code{TRUE}}

\item{...}{Curl options passed on to \code{\link[httr]{GET}} or \code{\link[httr]{POST}}}
}
\value{
either a data.frame or a list
}
\description{
Get repositories accessions
}
\examples{
\dontrun{
repositories_accessions(id = 1)
}
}

